즐겨찾기 isFavorite·selectFavlist·deleteFavoriteByOwner 이관(MemInfoDao→FavoriteDao)
분산된 즐겨찾기 로직 응집 마무리. 나머지 3개를 FavoriteDao로 옮기고 조회 접두어를 통일.

- MemInfoDao의 즐겨찾기 메서드 3개를 FavoriteDao로 이동(SQL·로직 동일)
  - getFavlist → selectFavlist (조회는 select* 접두어 통일, HashMap 반환이라 mapRow 비대상)
  - favDelete → deleteFavoriteByOwner (deleteFavorite(m_num,h_num)와 시그니처 동일해
    오버로드 불가 → 이름 구분. IDOR 소유자검증 2-인자(f_num,m_num) 의미 보존)
  - isFavorite (동사형 predicate라 이름 유지)
- MemInfoDao: 유일 사용처가 사라진 죽은 import java.util.HashMap 제거
- 호출처 갱신: checkfavorite·hugesodetail·favdelete·favlist.jsp
  (selectM_num은 MemInfoDao 유지, 즐겨찾기 호출만 FavoriteDao로 분리)
- hugesodetail.jsp: 사용처 0건이던 죽은 import favorite.model.FavoriteDto 제거
- 보안 가드(isLogin·isPost·checkCsrf·세션 m_num 강제·digitsOnly·escapeHtml) 전부 불변

javac EXIT=0.
