+++ src/main/webapp/mypage/favdelete.jsp
@@ -1,4 +1,5 @@
 <%@page import="meminfo.model.MemInfoDao"%>
+<%@page import="favorite.model.FavoriteDao"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <%
@@ -12,5 +13,5 @@ String f_num=util.SecurityUtil.digitsOnly(request.getParameter("f_num"));
 MemInfoDao dao=new MemInfoDao();
 // IDOR 방어: m_num을 세션 사용자로부터 도출해 본인 소유분만 삭제
 String m_num=dao.selectM_num(util.SecurityUtil.currentId(session));
-dao.favDelete(f_num, m_num);
+new FavoriteDao().deleteFavoriteByOwner(f_num, m_num);
 %>
