+++ src/main/webapp/mypage/favlist.jsp
@@ -1,6 +1,6 @@
 <%@page import="java.util.List"%>
 <%@page import="java.util.HashMap"%>
-<%@page import="meminfo.model.MemInfoDao"%>
+<%@page import="favorite.model.FavoriteDao"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <!DOCTYPE html>
@@ -108,8 +108,8 @@
 
 <%
 String m_id=(String)session.getAttribute("myid");
-MemInfoDao dao=new MemInfoDao();
-List<HashMap<String,String>> list=dao.getFavlist(m_id);
+FavoriteDao dao=new FavoriteDao();
+List<HashMap<String,String>> list=dao.selectFavlist(m_id);
 %>
 <div style="margin: 0 auto; width: 65%;height:50%; padding: 20px 20px 20px 20px; margin-top: 50px;">
 	<%--<h3><b>휴게소 즐겨찾기 목록</b></h3> --%>
